Tests: pyright fixes

+++ sinergym/utils/rewards.py
@@ -37,8 +37,8 @@ class LinearReward(BaseReward):
         self,
         temperature_variables: List[str],
         energy_variables: List[str],
-        range_comfort_winter: Tuple[int, int],
-        range_comfort_summer: Tuple[int, int],
+        range_comfort_winter: Tuple[float, float],
+        range_comfort_summer: Tuple[float, float],
         summer_start: Tuple[int, int] = (6, 1),
         summer_final: Tuple[int, int] = (9, 30),
         energy_weight: float = 0.5,
@@ -56,8 +56,8 @@ class LinearReward(BaseReward):
         Args:
             temperature_variables (List[str]): Name(s) of the temperature variable(s).
             energy_variables (List[str]): Name(s) of the energy/power variable(s).
-            range_comfort_winter (Tuple[int,int]): Temperature comfort range for cold season. Depends on environment you are using.
-            range_comfort_summer (Tuple[int,int]): Temperature comfort range for hot season. Depends on environment you are using.
+            range_comfort_winter (Tuple[float,float]): Temperature comfort range for cold season. Depends on environment you are using.
+            range_comfort_summer (Tuple[float,float]): Temperature comfort range for hot season. Depends on environment you are using.
             summer_start (Tuple[int,int]): Summer session tuple with month and day start. Defaults to (6,1).
             summer_final (Tuple[int,int]): Summer session tuple with month and day end. defaults to (9,30).
             energy_weight (float, optional): Weight given to the energy term. Defaults to 0.5.
@@ -190,8 +190,8 @@ class EnergyCostLinearReward(LinearReward):
         self,
         temperature_variables: List[str],
         energy_variables: List[str],
-        range_comfort_winter: Tuple[int, int],
-        range_comfort_summer: Tuple[int, int],
+        range_comfort_winter: Tuple[float, float],
+        range_comfort_summer: Tuple[float, float],
         energy_cost_variables: List[str],
         summer_start: Tuple[int, int] = (6, 1),
         summer_final: Tuple[int, int] = (9, 30),
@@ -212,8 +212,8 @@ class EnergyCostLinearReward(LinearReward):
         Args:
             temperature_variables (List[str]): Name(s) of the temperature variable(s).
             energy_variables (List[str]): Name(s) of the energy/power variable(s).
-            range_comfort_winter (Tuple[int,int]): Temperature comfort range for cold season. Depends on environment you are using.
-            range_comfort_summer (Tuple[int,int]): Temperature comfort range for hot season. Depends on environment you are using.
+            range_comfort_winter (Tuple[float,float]): Temperature comfort range for cold season. Depends on environment you are using.
+            range_comfort_summer (Tuple[float,float]): Temperature comfort range for hot season. Depends on environment you are using.
             summer_start (Tuple[int,int]): Summer s-sum(exp(violation)
                     for violation in temp_violations if violation > 0)ession tuple with month and day start. Defaults to (6,1).
             summer_final (Tuple[int,int]): Summer session tuple with month and day end. defaults to (9,30).
@@ -320,8 +320,8 @@ class ExpReward(LinearReward):
         self,
         temperature_variables: List[str],
         energy_variables: List[str],
-        range_comfort_winter: Tuple[int, int],
-        range_comfort_summer: Tuple[int, int],
+        range_comfort_winter: Tuple[float, float],
+        range_comfort_summer: Tuple[float, float],
         summer_start: Tuple[int, int] = (6, 1),
         summer_final: Tuple[int, int] = (9, 30),
         energy_weight: float = 0.5,
@@ -337,8 +337,8 @@ class ExpReward(LinearReward):
         Args:
             temperature_variables (List[str]): Name(s) of the temperature variable(s).
             energy_variables (List[str]): Name(s) of the energy/power variable(s).
-            range_comfort_winter (Tuple[int,int]): Temperature comfort range for cold season. Depends on environment you are using.
-            range_comfort_summer (Tuple[int,int]): Temperature comfort range for hot season. Depends on environment you are using.
+            range_comfort_winter (Tuple[float,float]): Temperature comfort range for cold season. Depends on environment you are using.
+            range_comfort_summer (Tuple[float,float]): Temperature comfort range for hot season. Depends on environment you are using.
             summer_start (Tuple[int,int]): Summer session tuple with month and day start. Defaults to (6,1).
             summer_final (Tuple[int,int]): Summer session tuple with month and day end. defaults to (9,30).
             energy_weight (float, optional): Weight given to the energy term. Defaults to 0.5.
@@ -403,8 +403,8 @@ class HourlyLinearReward(LinearReward):
         self,
         temperature_variables: List[str],
         energy_variables: List[str],
-        range_comfort_winter: Tuple[int, int],
-        range_comfort_summer: Tuple[int, int],
+        range_comfort_winter: Tuple[float, float],
+        range_comfort_summer: Tuple[float, float],
         summer_start: Tuple[int, int] = (6, 1),
         summer_final: Tuple[int, int] = (9, 30),
         default_energy_weight: float = 0.5,
@@ -418,8 +418,8 @@ class HourlyLinearReward(LinearReward):
         Args:
             temperature_variables (List[str]]): Name(s) of the temperature variable(s).
             energy_variables (List[str]): Name(s) of the energy/power variable(s).
-            range_comfort_winter (Tuple[int,int]): Temperature comfort range for cold season. Depends on environment you are using.
-            range_comfort_summer (Tuple[int,int]): Temperature comfort range for hot season. Depends on environment you are using.
+            range_comfort_winter (Tuple[float,float]): Temperature comfort range for cold season. Depends on environment you are using.
+            range_comfort_summer (Tuple[float,float]): Temperature comfort range for hot season. Depends on environment you are using.
             summer_start (Tuple[int,int]): Summer session tuple with month and day start. Defaults to (6,1).
             summer_final (Tuple[int,int]): Summer session tuple with month and day end. defaults to (9,30).
             default_energy_weight (float, optional): Default weight given to the energy term when thermal comfort is considered. Defaults to 0.5.
@@ -490,8 +490,8 @@ class NormalizedLinearReward(LinearReward):
         self,
         temperature_variables: List[str],
         energy_variables: List[str],
-        range_comfort_winter: Tuple[int, int],
-        range_comfort_summer: Tuple[int, int],
+        range_comfort_winter: Tuple[float, float],
+        range_comfort_summer: Tuple[float, float],
         summer_start: Tuple[int, int] = (6, 1),
         summer_final: Tuple[int, int] = (9, 30),
         energy_weight: float = 0.5,
@@ -504,8 +504,8 @@ class NormalizedLinearReward(LinearReward):
         Args:
             temperature_variables (List[str]]): Name(s) of the temperature variable(s).
             energy_variables (List[str]): Name(s) of the energy/power variable(s).
-            range_comfort_winter (Tuple[int,int]): Temperature comfort range for cold season. Depends on environment you are using.
-            range_comfort_summer (Tuple[int,int]): Temperature comfort range for hot season. Depends on environment you are using.
+            range_comfort_winter (Tuple[float,float]): Temperature comfort range for cold season. Depends on environment you are using.
+            range_comfort_summer (Tuple[float,float]): Temperature comfort range for hot season. Depends on environment you are using.
             summer_start (Tuple[int,int]): Summer session tuple with month and day start. Defaults to (6,1).
             summer_final (Tuple[int,int]): Summer session tuple with month and day end. defaults to (9,30).
             energy_weight (float, optional): Default weight given to the energy term when thermal comfort is considered. Defaults to 0.5.
